requirements.txt
flask
line-bot-sdk>=3.0.0
googletrans==4.0.0-rc1
gtts==2.5.0
gunicorn
